fix(permissions): read power levels as Long, render the ceiling as ∞
A power level is bounded only by the canonical-JSON integer range (±2^53),
and servers use the top of it: Conduit gives a room-version-12 creator
9007199254740990 in `users`, and marks state events nobody may send
9007199254740991. parsePowerLevels read them with optInt, which truncates
rather than clamps — 2^53-2 has low 32 bits 0xFFFFFFFE, so the creator's
level came back as -2. Below users_default, that made the account owning
the room its least privileged member: no composer, no pin, no moderation,
and last in the member list.

PowerLevelsInfo is Long throughout, parsed with optLong, and
CREATOR_POWER_LEVEL is Long.MAX_VALUE — Int.MAX_VALUE no longer sits above
every level a room can express. The power-levels narrator had the same bug
in cosmetic form: two distinct huge levels truncated to the same Int and
narrated as no change.

Ceiling levels are a server spelling "unreachable", so formatPowerLevel
collapses anything from 2^53-2 up to ∞ for display — the member badge (which
already did this for creators), the power-levels dialog and the narrator.
Comparisons still use the real values.

+++ app/src/main/java/net/vrkknn/andromuks/RoomItem.kt
@@ -155,18 +155,25 @@ data class RoomState(
 @Immutable
 data class ServerAclInfo(val allow: List<String>, val deny: List<String>, val allowIpLiterals: Boolean)
 
-/** Power levels information for a room */
+/**
+ * Power levels information for a room.
+ *
+ * Every level is a **Long**, not an Int. Matrix bounds power levels only by the JSON canonical
+ * integer range (±2^53), and servers do use the top of it: Conduit hands a room-version-12 creator
+ * `9007199254740990` in `users`. Parsing that as an Int truncates it to -2 — the creator came out
+ * ranked *below* everyone.
+ */
 @Immutable
 data class PowerLevelsInfo(
-    val users: Map<String, Int>,
-    val usersDefault: Int,
-    val redact: Int,
-    val kick: Int = 50, // Default kick power level
-    val ban: Int = 50, // Default ban power level
-    val invite: Int = 50, // Default invite power level (used to accept knocks)
-    val events: Map<String, Int> = emptyMap(),
-    val eventsDefault: Int = 0,
-    val stateDefault: Int = 50, // Per Matrix spec: default PL for state events not in events map
+    val users: Map<String, Long>,
+    val usersDefault: Long,
+    val redact: Long,
+    val kick: Long = 50, // Default kick power level
+    val ban: Long = 50, // Default ban power level
+    val invite: Long = 50, // Default invite power level (used to accept knocks)
+    val events: Map<String, Long> = emptyMap(),
+    val eventsDefault: Long = 0,
+    val stateDefault: Long = 50, // Per Matrix spec: default PL for state events not in events map
 )
 
 @Immutable
